+++ package.json
@@ -2,7 +2,7 @@
   "name": "electron-acrylic-window",
   "version": "0.4.7",
   "description": "Add vibrancy effect for electron",
-  "main": "index.js",
+  "main": "src/index.js",
   "keywords": [
     "node",
     "vibrancy",
