+++ tsconfig.json
@@ -4,10 +4,12 @@
 		"target": "es5",
 		"lib": [ "es5" ],
 		"sourceMap": true,
-		"outDir": "./",
-		"strict": true
+		"declaration": true,
+		"strict": true,
+		"rootDir": "src"
 	},
 	"exclude": [
-		"node_modules"
+		"node_modules",
+		"test"
 	]
 }
